Rename interval::interval --> interval::base
