+++ src/backend/pitchbend.rs
@@ -8,7 +8,7 @@ use midi_msg::{Channel, ChannelModeMsg, ChannelVoiceMsg, ControlChange, MidiMsg}
 use crate::{
     backend::r#trait::BackendState,
     config::r#trait::Config,
-    interval::{interval::Semitones, stacktype::r#trait::StackType},
+    interval::{base::Semitones, stacktype::r#trait::StackType},
     msg,
 };
