+++ src/interval/base.rs
