+++ src/interval/mod.rs
@@ -1,4 +1,4 @@
-pub mod interval;
+pub mod base;
 pub mod stack;
 pub mod stacktype;
 pub mod temperament;
