+++ src/tui/grid.rs
@@ -10,7 +10,7 @@ use ratatui::{
 use crate::{
     config::r#trait::Config,
     interval::{
-        interval::Semitones,
+        base::Semitones,
         stack::Stack,
         stacktype::r#trait::{FiveLimitStackType, PeriodicStackType, StackCoeff, StackType},
     },
